bricks, Route hacked to support exception rendering
Routes now have an optional argument called httpexception_handlers
which is a map from an HTTPException type to a component that will
handle it if it occurs at that route.

create_app must manually load handlers for each route.

routes are no longer added to MainComponent dependencies, only
the root, the rest are loaded manually by create_app.

✗untested
✗non-ideal implementation

please review